Bug 817762: bulletproofify the timeout handlers; r=mcote
This runs the handlers in a thread, so that we can log when they take too long.
This also catches exceptions just about everywhere, and retries when necessary.

I've tested this with:
 - failed MySQL server
 - slow MySQL server (locked table)
 - bad schema (devices table missing)

In all cases, not much happens during the failure condition, but once the
situation is rectified, the deamon gets right back to work processing timeouts
and whatnot.